+++ R/box_data.R
@@ -1,0 +1,26 @@
+#' @title Box data for box_plot, include Value, Level1 and Level2 columns.
+#' @description Box data for box_plot, include Value, Level1 and Level2 columns.
+#' @author benben-miao
+#'
+#' @docType data
+#' @format Dataframe: include Value, Level1 and Level2 columns.
+#' @usage data(box_data)
+#'
+#' @keywords datasets
+#'
+#' @references https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/BoxStat/
+#'
+#' @examples
+#' # 1. Library TOmicsVis package
+#' library(TOmicsVis)
+#'
+#' # 2. Load example data
+#' data(box_data)
+#'
+#' # 3. View example data
+#' box_data
+#'
+"box_data"
+
+# box_data <- data
+# usethis::use_data(box_data)